added base user view and removed unwanted activities
