+++ D_Plan/app/src/main/AndroidManifest.xml
@@ -19,9 +19,6 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".ToDoActivity"
-            android:label="@string/app_name" />
-        <activity
             android:name=".Login"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.DayNight.NoActionBar">
@@ -43,7 +40,11 @@
             android:name=".Add_Disaster"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.DayNight.NoActionBar" />
-        <activity android:name=".Delete_record"
+        <activity
+            android:name=".Delete_record"
+            android:screenOrientation="portrait"
+            android:theme="@style/Theme.AppCompat.DayNight.NoActionBar" />
+        <activity android:name=".User_awareness"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.DayNight.NoActionBar"/>
     </application>
