+++ D_Plan/app/src/main/res/layout/activity_user_awareness.xml
