+++ D_Plan/app/src/main/res/menu/user_menu.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:id="@+id/menu_refresh"
+        android:orderInCategory="101"
+        android:title="@string/menu_refresh"
+        android:icon="@drawable/refresh"
+        app:showAsAction="ifRoom">
+    </item>
+    <item android:id="@+id/logout"
+        android:orderInCategory="101"
+        android:title="Logout"
+        android:icon="@drawable/logout"
+        app:showAsAction="ifRoom">
+    </item>
+</menu>
